Update route names to match current hostnames
At the moment the cover API uses the hostname cover.dandigbib.org while
the upload API uses the hostname upload.cover.dandigbib.org.

To match this on Platform.sh we will use cover.dandigbib.org as the
project domain. This will make the cover API the default hostname
and the upload API a subdomain of this.